例文追加画面　例文追加セクション周り型大幅変更
