+++ frontend-next/src/pages/englishBot/addExample.tsx
@@ -1,45 +1,24 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Container } from '@mui/material';
 import { Layout } from '@/components/templates/layout/Layout';
 import { Title } from '@/components/ui-elements/title/Title';
 import { AddExampleSection } from '@/components/ui-forms/englishbot/addExample/addExampleSection/AddExampleSection';
 import { messageState } from '@/atoms/Message';
 import { useSetRecoilState } from 'recoil';
-import { Card } from '@/components/ui-elements/card/Card';
-import { Button } from '@/components/ui-elements/button/Button';
-import { submitExampleSentenseAPI } from '@/api/englishbot/submitExampleSentenseAPI';
 import { AssociateExampleandWordSection } from '@/components/ui-forms/englishbot/addExample/searchRelatedWordSection/AssociateExampleandWordSection';
-
-// TODO 共通libに持っていく
-export type InputExampleData = {
-  wordName?: string;
-  exampleJa?: string;
-  exampleEn?: string;
-};
 
 type Props = {
   isMock?: boolean;
 };
 
 export default function EnglishBotAddExamplePage({ isMock }: Props) {
-  // TODO ↓のステート　コンポーネント内に入れる Button含めて　Cardからコンポーネントに入れる
-  const [inputExampleData, setInputExampleData] = useState<InputExampleData>({});
   const setMessage = useSetRecoilState(messageState);
 
   const contents = () => {
     return (
       <Container>
         <Title label="Add Example Sentense"></Title>
-        <Card variant="outlined" attr="margin-vertical" header="例文追加">
-          <AddExampleSection inputExampleData={inputExampleData} setInputExampleData={setInputExampleData} />
-          <Button
-            label={'登録'}
-            attr={'after-inline'}
-            variant="contained"
-            color="primary"
-            onClick={(e) => submitExampleSentenseAPI({ inputExampleData, setMessage, setInputExampleData })}
-          />
-        </Card>
+        <AddExampleSection setMessage={setMessage} />
         <AssociateExampleandWordSection setMessage={setMessage} />
       </Container>
     );
