+++ quizzer-lib/src/english/example/addExample/dto.ts
@@ -1,0 +1,6 @@
+// 例文追加APIリクエスト型
+export interface AddExampleAPIRequestDto {
+  exampleEn: string
+  exampleJa: string
+  wordName: string
+}
