+++ quizzer-lib/src/english/index.ts
@@ -1,1 +1,2 @@
 export * from './word'
+export * from './example'
