+++ quizzer-lib/src/interfaces/api/english/request.ts
@@ -1,10 +1,3 @@
-// 例文追加APIリクエスト型
-export interface AddExampleAPIRequestDto {
-  exampleEn: string
-  exampleJa: string
-  wordName: string
-}
-
 //
 export interface ChangeAssociationOfExampleRequestDto {
   wordName: string
